Changed Monte Carlo Tree Search
Below are some of the changed I made:
explore_exploit_selection():
changed how score is calculated to the upper confidence bounad equation. The originial implementation of epsilon-greedy-search would work as well and we could use it as one of things test to see which way works better.
update_score():
created a function in the node class to check which index the agent is before adding the new score to make the ghosts and agent have an oppsite score and wins to make sure that traversals of the tree pick the ghost's best option.
backpropagate():
uses the new update_score() function
state_heuistic():
calculates the heuristic of a non-terminal state. Currently just the current score minus a fourth of the L1 Norm distance to the nearest food. This heuristic has problems since Pacman wastes several turns grabing the last food.
Simulate():
Only simulates 2 moves for each player currently. We can look into increasing it and seeing how the results change for the paper. Then, if the state isn't terminal, it uses the state_heurstic() function.