Refactored Monte Carlo Tree Search
Added a global variable in the class to save the previous tree. Then it uses the next time the class is called to look for the node in the tree that represents that state and starts with this sub-tree instead of from scratch. We can examine both.

best_score_selection():
Made it based on the win-rate also so that actions will no win rate can't be chosen.

Class node():
Made sure children nodes have an updates agent index to their parent.